+++ desktop_version/src/editor.cpp
@@ -512,7 +512,7 @@ void editorclass::getlin(const enum textmode mode, const std::string& prompt, st
     ed.oldenttext = key.keybuffer;
 }
 
-const int* editorclass::loadlevel( int rxi, int ryi )
+const short* editorclass::loadlevel( int rxi, int ryi )
 {
     //Set up our buffer array to be picked up by mapclass
     rxi -= 100;
@@ -522,7 +522,7 @@ const int* editorclass::loadlevel( int rxi, int ryi )
     if(rxi>=mapwidth)rxi-=mapwidth;
     if(ryi>=mapheight)ryi-=mapheight;
 
-    static int result[1200];
+    static short result[1200];
 
     for (int j = 0; j < 30; j++)
     {
