+++ desktop_version/src/editor.h
@@ -130,7 +130,7 @@ class editorclass{
 
   void reset();
   void getlin(const enum textmode mode, const std::string& prompt, std::string* ptr);
-  const int* loadlevel(int rxi, int ryi);
+  const short* loadlevel(int rxi, int ryi);
 
   void placetile(int x, int y, int t);
 
@@ -190,7 +190,7 @@ class editorclass{
   std::vector<std::string> getLevelDirFileNames( );
   static const int maxwidth = 20, maxheight = 20; //Special; the physical max the engine allows
   static const int numrooms = maxwidth * maxheight;
-  int contents[40 * 30 * numrooms];
+  short contents[40 * 30 * numrooms];
   int vmult[30 * maxheight];
   int numtrinkets();
   int numcrewmates();
